Automate person-directory configuration (#1883)
* Updated config

* Updated config

* Updated config

* Updated config

* removed old ehcache dependency

* Updated travis build

* Updated config
