+++ build.gradle
@@ -36,7 +36,7 @@ def projectMavenDeployer = {
         inceptionYear project.projectInceptionYear
         name project.name
         description project.name
-        
+
         issueManagement {
             system project.projectIssueSystem
             url project.projectIssueUrl
@@ -226,6 +226,12 @@ def projectMavenDeployer = {
 }
 
 apply from: rootProject.file("gradle/dependencies.gradle")
+apply plugin: "com.gradle.build-scan"
+
+buildScan {
+    licenseAgreementUrl = "https://gradle.com/terms-of-service"
+    licenseAgree = "yes"
+}
 
 buildscript {
     repositories {
@@ -243,6 +249,7 @@ buildscript {
         classpath "com.netflix.nebula:gradle-lint-plugin:0.27.0"
         classpath "org.standardout:gradle-versioneye-plugin:1.4.0"
         classpath "org.springframework.build.gradle:propdeps-plugin:0.0.7"
+        classpath "com.gradle:build-scan-plugin:1.0"
     }
 }
 
@@ -284,10 +291,11 @@ subprojects {
     apply plugin: "propdeps-idea"
     apply plugin: "propdeps-eclipse"
 
+
     apply from: rootProject.file("gradle/tasks.gradle")
 
     ext.libraries = rootProject.ext.libraries;
-    
+
     repositories {
         mavenLocal()
         mavenCentral()
