+++ cas-server-core-authentication/build.gradle
@@ -11,6 +11,7 @@ dependencies {
     compile libraries.spring
     compile libraries.hibernate
     compile libraries.springcloud
+    compile libraries.guava
     testCompile project(":cas-server-core-web")
     compile libraries.springsecurity
 }
