+++ cas-server-core-configuration/src/main/java/org/apereo/cas/configuration/model/core/authentication/AuthenticationProperties.java
@@ -27,9 +27,7 @@ import org.apereo.cas.configuration.model.support.x509.X509Properties;
 import org.springframework.boot.context.properties.NestedConfigurationProperty;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 /**
  * This is {@link AuthenticationProperties}.
@@ -39,20 +37,21 @@ import java.util.Map;
  */
 public class AuthenticationProperties {
 
-    private Map<String, String> attributes = new HashMap();
+    @NestedConfigurationProperty
+    private PrincipalAttributesProperties attributeRepository = new PrincipalAttributesProperties();
 
     @NestedConfigurationProperty
     private RestAuthenticationProperties rest = new RestAuthenticationProperties();
-    
+
     @NestedConfigurationProperty
     private List<LdapAuthenticationProperties> ldap = new ArrayList<>();
-    
+
     @NestedConfigurationProperty
     private ThrottleProperties throttle = new ThrottleProperties();
-    
+
     @NestedConfigurationProperty
     private SamlIdPProperties samlIdp = new SamlIdPProperties();
-    
+
     @NestedConfigurationProperty
     private AuthenticationExceptionsProperties exceptions =
             new AuthenticationExceptionsProperties();
@@ -60,7 +59,7 @@ public class AuthenticationProperties {
     @NestedConfigurationProperty
     private AuthenticationPolicyProperties policy =
             new AuthenticationPolicyProperties();
-    
+
     @NestedConfigurationProperty
     private PasswordPolicyProperties passwordPolicy = new PasswordPolicyProperties();
 
@@ -81,7 +80,7 @@ public class AuthenticationProperties {
 
     @NestedConfigurationProperty
     private TrustedAuthenticationProperties trusted = new TrustedAuthenticationProperties();
-    
+
     @NestedConfigurationProperty
     private JaasAuthenticationProperties jaas = new JaasAuthenticationProperties();
 
@@ -90,7 +89,7 @@ public class AuthenticationProperties {
 
     @NestedConfigurationProperty
     private MultifactorAuthenticationProperties mfa = new MultifactorAuthenticationProperties();
-    
+
     @NestedConfigurationProperty
     private MongoAuthenticationProperties mongo = new MongoAuthenticationProperties();
 
@@ -123,7 +122,7 @@ public class AuthenticationProperties {
 
     @NestedConfigurationProperty
     private X509Properties x509 = new X509Properties();
-    
+
     public AuthenticationExceptionsProperties getExceptions() {
         return exceptions;
     }
@@ -139,7 +138,7 @@ public class AuthenticationProperties {
     public void setPolicy(final AuthenticationPolicyProperties policy) {
         this.policy = policy;
     }
-        
+
     public PasswordPolicyProperties getPasswordPolicy() {
         return passwordPolicy;
     }
@@ -324,13 +323,6 @@ public class AuthenticationProperties {
         this.trusted = trusted;
     }
 
-    public Map<String, String> getAttributes() {
-        return attributes;
-    }
-
-    public void setAttributes(final Map<String, String> attributes) {
-        this.attributes = attributes;
-    }
 
     public List<LdapAuthenticationProperties> getLdap() {
         return ldap;
@@ -346,5 +338,13 @@ public class AuthenticationProperties {
 
     public void setRest(final RestAuthenticationProperties rest) {
         this.rest = rest;
+    }
+
+    public PrincipalAttributesProperties getAttributeRepository() {
+        return attributeRepository;
+    }
+
+    public void setAttributeRepository(final PrincipalAttributesProperties attributeRepository) {
+        this.attributeRepository = attributeRepository;
     }
 }
