+++ cas-server-webapp/src/main/resources/application.properties
@@ -7,7 +7,7 @@ server.port=8443
 server.ssl.key-store=file:/etc/cas/thekeystore
 server.ssl.key-store-password=changeit
 server.ssl.key-password=changeit
-#server.ssl.ciphers= 
+#server.ssl.ciphers=
 #server.ssl.client-auth=
 #server.ssl.enabled=
 #server.ssl.key-alias=
@@ -54,7 +54,6 @@ spring.cloud.bus.ack.enabled=true
 ##
 # CAS Admin Status Endpoints
 #
-cas.securityContext.adminpages.ip=127\.0\.0\.1
 endpoints.enabled=true
 endpoints.sensitive=true
 management.context-path=/status
@@ -94,10 +93,10 @@ cas.authn.accept.users=casuser::Mellon
 ##
 # CAS Authentication Attributes
 #
-cas.authn.attributes.uid=uid
-cas.authn.attributes.displayName=displayName
-cas.authn.attributes.cn=commonName
-cas.authn.attributes.affiliation=groupMembership
+cas.authn.attributeRepository.attributes.uid=uid
+cas.authn.attributeRepository.attributes.displayName=displayName
+cas.authn.attributeRepository.attributes.cn=commonName
+cas.authn.attributeRepository.attributes.affiliation=groupMembership
 
 ##
 # CAS Groovy Shell Console
