+++ gradle/dependencies.gradle
@@ -101,7 +101,36 @@ ext.libraries = [
                     exclude(group: "org.javassist", module: "javassist")
                     force = true
                 },
-                dependencies.create("org.javassist:javassist:$javassistVersion")
+                dependencies.create("org.javassist:javassist:$javassistVersion"),
+                dependencies.create("org.springmodules:spring-modules-cache:0.8") {
+                    exclude(group: "commons-logging", module: "commons-logging")
+                    exclude(group: "ant", module: "ant")
+                    exclude(group: "commons-lang", module: "commons-lang")
+                    exclude(group: "xpp3", module: "xpp3_min")
+                    exclude(group: "jcs", module: "jcs")
+                    exclude(group: "jboss", module: "jboss-system")
+                    exclude(group: "jboss", module: "javassist")
+                    exclude(group: "jboss", module: "jboss-cache")
+                    exclude(group: "jboss", module: "jboss-jmx")
+                    exclude(group: "jboss", module: "jboss-minimal")
+                    exclude(group: "jboss", module: "jboss-common")
+                    exclude(group: "jgroups", module: "jgroups-all")
+                    exclude(group: "log4j", module: "log4j")
+                    exclude(group: "xjavadoc", module: "xjavadoc")
+                    exclude(group: "jini", module: "xjavadoc")
+                    exclude(group: "jini", module: "jsk-platform")
+                    exclude(group: "jini", module: "boot")
+                    exclude(group: "jini", module: "jsk-lib")
+                    exclude(group: "jini", module: "mahalo")
+                    exclude(group: "jini", module: "reggie")
+                    exclude(group: "jini", module: "start")
+                    exclude(group: "jini", module: "webster")
+                    exclude(group: "ehcache", module: "ehcache")
+                    exclude(group: "gigaspaces", module: "gigaspaces-ce")
+                    exclude(group: "commons-collections", module: "commons-collections")
+                    exclude(group: "org.springframework", module: "spring")
+                    force = true
+                }
         ],
         hazelcast               : dependencies.create("com.hazelcast:hazelcast:$hazelcastVersion"),
         hibernate               : [
